escrowapp/backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-dotenv
pydantic[email]
python-multipart
email-validator
